Added ressource view for associations
