+++ templates/file/association_file_tree.html
@@ -32,7 +32,7 @@
             {% if is_root %}
                 <tr>
                     <td><i class="fa fa-folder"></i> <a
-                            href="{% url 'resource-file-tree' association.id %}">Ressources</a></td>
+                            href="{% url 'association-resource-tree' association.id %}">Ressources</a></td>
                 </tr>
             {% endif %}
         {% endif %}
